Implement Codable on JSON and JSONObject
